+++ pkg/phlaredb/symdb/resolver_tree_test.go
@@ -1,0 +1,43 @@
+package symdb
+
+import (
+	"context"
+	"testing"
+
+	"github.com/stretchr/testify/require"
+)
+
+func Test_memory_Resolver_ResolveTree(t *testing.T) {
+	s := newMemSuite(t, [][]string{{"testdata/profile.pb.gz"}})
+	expectedFingerprint := pprofFingerprint(s.profiles[0], 0)
+	r := NewResolver(context.Background(), s.db)
+	defer r.Release()
+	r.AddSamples(0, s.indexed[0][0].Samples)
+	resolved, err := r.Tree()
+	require.NoError(t, err)
+	require.Equal(t, expectedFingerprint, treeFingerprint(resolved))
+}
+
+func Test_block_Resolver_ResolveTree(t *testing.T) {
+	s := newBlockSuite(t, [][]string{{"testdata/profile.pb.gz"}})
+	defer s.teardown()
+	expectedFingerprint := pprofFingerprint(s.profiles[0], 1)
+	r := NewResolver(context.Background(), s.reader)
+	defer r.Release()
+	r.AddSamples(0, s.indexed[0][1].Samples)
+	resolved, err := r.Tree()
+	require.NoError(t, err)
+	require.Equal(t, expectedFingerprint, treeFingerprint(resolved))
+}
+
+func Benchmark_block_Resolver_ResolveTree(t *testing.B) {
+	s := newBlockSuite(t, [][]string{{"testdata/big-profile.pb.gz"}})
+	defer s.teardown()
+	t.ResetTimer()
+	t.ReportAllocs()
+	for i := 0; i < t.N; i++ {
+		r := NewResolver(context.Background(), s.reader)
+		r.AddSamples(0, s.indexed[0][0].Samples)
+		_, _ = r.Tree()
+	}
+}
